Move appshell and footer components into shared file
